requirements.txt
Django==3.2.7
mysqlclient==2.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
pyyaml==6.0
django-rest-auth
django-allauth
django-braces

